+++ app/Http/Controllers/PostController.php
@@ -10,7 +10,7 @@ class PostController extends Controller
 {
     public function index(){
         $posts = Post::orderBy('id','desc')
-        ->with(['categories','comments'])
+        ->with(['categories','comments','user'])
         ->paginate(10);
 
         return response()->json($posts);
@@ -18,14 +18,18 @@ class PostController extends Controller
 
     public function create(Request $request){
         $params = $request->only(['subject','content']);
+        $params['user_id']=$request->user()->id;
         $post = Post::create($params);
         $ids = $request->input('category_ids');
         $post->categories()->sync($ids);
-        return response()->json($post);
+        $result = Post::where('id',$post->id)
+        ->with(['user','categories'])->first();
+        return response()->json($result);
     }
 
     public function read($id){
-        $post = Post::where('id',$id)->with('comments')->first();
+        $post = Post::where('id',$id)
+        ->with(['comments','user'])->first();
         //$post = Post::find($id);
 
         if(!$post){
@@ -35,10 +39,17 @@ class PostController extends Controller
     }
 
     public function update(Request $request,$id){
+
         $post = Post::find($id);
         if(!$post){
             return response()->json(['message'=>'조회한 데이터가 없습니다.'],404);
         }
+
+        $user = $request->user();
+        if($user->id!==$post->user_id){
+            return response()->json(['message'=>'권한이 없습니다.'],403);
+        }
+
         $subject = $request->input('subject');
         $content = $request->input('content');
         $ids = $request->input('category_ids');
@@ -51,11 +62,18 @@ class PostController extends Controller
         return response()->json($post);
     }
 
-    public function delete($id){
+    public function delete(Request $request, $id){
         $post = Post::find($id);
         if(!$post){
             return response()->json(['message'=>'조회한 데이터가 없습니다.'],404);
         }
+
+        $user = $request->user();
+        if($user->id!==$post->user_id){
+            return response()->json(['message'=>'권한이 없습니다.'],403);
+        }
+
+
         $post->delete();
 
         return response()->json(["message"=>'delete one']);
